Updated releng/autorelease Project: controller  e6215c91f474affe19ac2e8aab11983d75ead9b2
BUG-2953 : Unable to read from datastore root with clustering enabled

This patch adds a NormalizedNodeAggregator which makes it possible to read
from the root of the datastore tree.

Writing, merging and checking for the existence of the root path is not
supported yet but we could in a future patches if needed.

Change-Id: I892bdc733ec316f41926d8bbccdbbdc6c31a5c25
Signed-off-by: Moiz Raja <moraja@cisco.com>