Updated releng/autorelease Project: controller  e1e2b8993826feb599a77a3464d8f46a6c7897ca
Add new DeleteEntries class wih long "fromIndex" field

Added a new DeleteEntries message class that defines the fromIndex field
as a long. The previous DeleteEntries class (inner class of RaftActor)
defined it as an int which could cause rollover. The previous DeleteEntries
must be kept for the time being for backwards compatibility as instances
are persisted.

Change-Id: I774128d9a1af2410a23d2739c37484a508c8a80f
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>