Updated releng/autorelease Project: tsdr  77ef5c3ef6bcfeac8fe4bfb6d04bd5b80a318d52
Introduce AbstractList command to display the list of metrics in SPI
Made changes in persistence side to allow the list command

Note: The list command is giving issue with the current version of H2
and needs to be further investigated wherein it cannot take the where clause
and currently H2 list command list the last 1000 metrics. Im pushing this one to not hold on SPI changes done.

Change-Id: I5b58baa2dccdd6d6aab3f63c54379cf86a2daada
Signed-off-by: Basheeruddin Ahmed <syedbahm@cisco.com>