Updated releng/autorelease Project: yangtools  8a182d66729717ab0308292da5986ef4191d8bd7
Add a revisionless QName constructor

We already provide a string-based constructor, but that requires a valid
revision date. A QNameModule does not require a revision, which is
reflected by other factory methods. Add a two-string factory method,
which does not take a revision argument.

Change-Id: I7c9bd36fc972a843b6b45ef972400a8f77f9c9b0
Signed-off-by: Robert Varga <rovarga@cisco.com>

Fix DataTreeCandidateNode API definition

The javadocs for getDataBefore() and getDataAfter() were switched. Also
add @Nonnull annotations to express the intent to users.

Change-Id: I0957366830f43f0c02e69ae7bddfabe018955bef
Signed-off-by: Robert Varga <rovarga@cisco.com>