Updated releng/autorelease Project: yangtools  1570aee2c8961d093b676f916f93b5ed789b4376
Expose YangInstanceIdentifier.EMPTY

Quite a few users have custom-made versions of this field. Expose it as
an API element.

Change-Id: I2e77b6bc7751b8e4be92df139c9b1cf782328414
Signed-off-by: Robert Varga <rovarga@cisco.com>

SchemaAwareApplyOperation.checkDeleteApplicable() can be static

This method does not touch local state, so make it static.

Change-Id: I973fb25e4989c7969ddb5df18e2d683d2cbc15ef
Signed-off-by: Robert Varga <rovarga@cisco.com>

Dump the modification which fails to produce a root node

For CDS implementation, there seems to be a bug where we fail to create
an appropriate changeset. Also be sure to send debug on how the
candidate is being applied.

Change-Id: Ie0eccd831169ff42c002851633926713dba20c74
Signed-off-by: Robert Varga <rovarga@cisco.com>