Updated releng/autorelease Project: controller  5f553b1657b97adbeb59b8e346d6eed8148b281b
Reduce non-existent Tx chain logging on close in ShardDataTree

In ShardDataTree#closeTransactionChain, if the Tx chain doesn't exist it
emits a warning. However the CloseTransactionChain message is
broadcasted to all shards so there will always be a non-existent chain
on some shards so I reduced the log level to debug.

Change-Id: Ia4243c4a2987fa4f43d272a7435a9d7b5d43445b
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>