Updated releng/autorelease Project: controller  d1d24d3742ffcda5e16a1d35e15a5627d5eb05f9
Fix ReplicationAndSnapshotsIntegrationTest failure

Saw this fail sproradically a few times on jenkins with
InvalidActorNameException trying to kill and re-create the actor with
the same name. We wait for the actor to terminate but it seems it may
not be removed from akka at that point. So I added a retry loop for
creating the actor.

Change-Id: Ia2d7388b00df329d62a4683a18f49067cfdc17a0
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>