Updated releng/autorelease Project: nic  e654dcc9e4d77d9e37f0c86c0893d5e5b383e853
update the intent yang to introduce primitive actions

longer version: between the combination of yang and yangtools
it is not possible to make the system really do what i wanted
so this is a compromise. The big change is that to deal with
"derrived" entity types with values we have to move away from
using identities and instead use groupings. But in order to
get yangtools to generate builder classes those groupings
have to be referenced in a container. So, to get things to
work, i had to create the intent with a bunch of choices.

I did split subject, conditions, etc. out into separate files
to get some modularity.

Change-Id: I62e458458bd815eddffc966dc3bf059b85e70cd9
Signed-off-by: David K. Bainbridge <dbainbri@ciena.com>