Updated releng/autorelease Project: integration  8aee954b99437954936ac2db30dfed495b091257
new statistic manager tests

The idea of these tests is that flows are spread over the switches and
their tables the way we want. Eighter we want to have corner cases (all
flows on one switch in one table) or we want to have flows spread over
switches and tables according some rules.

The library generates flow details (a list) which are then used for
(de)configuration and checking the operational inventory datastore.

Testplan is not created yet. I'll do it next week.

Change-Id: I9f0ffa6f05c22f2aa75fc2cc2120f0390086daa3
Signed-off-by: Peter Gubka <pgubka@cisco.com>