Updated releng/autorelease Project: integration  8a94c4f376a9bedb9270d1ff256af7d91ee80167
Tcpmd5user suite: Use pcc-mock in reconnecting mode

The initial specification for TCPMD5 user-facing feature
system tests states that pcc-mock should be used in reconnecting mode.

Pcc-mock already supports --reconnect command line argument,
and some of the previous failures can be attributed to the fact
that this argument was not used in initial check-in of tcpmd5user suite.

This change makes pcc-mock reconnecting in 1 second intervals.

Also, test for not seeing prompt is more narrow now.

Change-Id: I07c36d111908aa0f70a3741f11e897d310aaf59c
Signed-off-by: Vratko Polak <vrpolak@cisco.com>