Updated releng/autorelease Project: integration  0d7b49b8cbe588d60a3155920c592b2d4011e436
Add odl-dsbenchmark-impl feature to Integration distribution

The coretutorials has not declared any feature to be user-facing,
but Dsbenchmark can be useful for users (to compare installations)
and it is very useful for CI: tracking DS performance.

As CI only uses builds from ${project}-distribution-${stream} jobs,
this change is needed for CI tests to happen.

Change-Id: If08d6893f0434d1eae2e47aed389f186588e9d30
Signed-off-by: Vratko Polak <vrpolak@cisco.com>