Updated releng/autorelease Project: packetcable  30298f4f3110a14d75dfcab37e11b70a6ce9a1b0
The second patch of an estimated 4 to complete the COPS message refactoring as described in Bugzilla report #2943. This patch completes the refactoring for all of the objects that can be contained within a COPS message. Although this changelist is quite large, it was not possible to subdivide it any further as it was necessary to refactor each of the classes being extended by COPSObjectBase. Most of the changes made at the service level are to the COPS object construction; however, a few fixes have also been made to address some of the compiler warnings. This patch also should have taken out all of the DOS CR characters in any of the source being altered at this time.

Change-Id: I299f2360dac8455b2f4e87d9e8e751ae7c843f3b
Signed-off-by: Steven Pisarski <s.pisarski@cablelabs.com>