Updated releng/autorelease Project: packetcable  830a47e093230c070a7430aad16bde4520313e7d
Completed COPS Message refactoring. Was planning on one additional patch starting with the COPSHeader; however, the dependencies of COPSHeader were everywhere the COPSMsg objects were being instantiated and determined it would be easier merging this as the final patch versus splitting these changes up. Please see Bugzilla report #2943 for more details. In addition to the restructuring of the COPSMsg classes for immutability, semantics by adding in enumerations, and testability, this patch is also fixing some of the code warnings and removing MS-DOS CR characters.

Change-Id: Ica1e55209728fab5c71d0bffb52eb4ca0401e4dc
Signed-off-by: Steven Pisarski <s.pisarski@cablelabs.com>