Updated releng/autorelease Project: packetcable  f7f98d84abce359950f2bb79a2ccce7776b474b0
Created a client-type enumeration as part of the original COPS refactor and have changed it back to an short value due to a misinterpretation of the COPS specifications. Additionally changed the enumeration held by the COPSDecision to not leverage the ordinal value as a mapping between the COPS payload which comes in as two bytes.

Change-Id: I2c86fd4fc134afa12de1fee12a52279f1a9fdf31
Signed-off-by: Steven Pisarski <s.pisarski@cablelabs.com>