Updated releng/autorelease Project: controller  0dd3756434941ff5635a3a3bc60a2f0b59338c1c
Bug 3161: Remove serialVersionUID to maintain backwards compatibility

The serialVersionUID field was added earlier in Lithium for
RaftActor.DeleteEntries and RaftActor.UpdateElectionTerm however it
breaks backwards compatibility. We need to remove serialVersionUID.

DeleteEntries has already been deprecated for a new class that has
serialVersionUID set and is not an inner class. I will do that same for
UpdateElectionTerm in another patch.

Change-Id: I8180de5865d89d3fea190e48ebeb825743af959d
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>