Updated releng/autorelease Project: ovsdb  77af48ad9d278b2a7e6d54442bd23bff313e08ee
Client should be added to cache before we push the
info(OvsdbNodeAugmentation) to operational data store.
With the existing implementation we were putting task
in invoker queue to dump the data to operational data
store and then registtering for the connection callback.
After that it's adding client to cache. So if invoker
writes the data to operational data store before
it adds client to cache, any code listening to
OvsdbNodeAugmentaiton, might not be able to do any
operation if that key is not there in connection
manager cache. It works if invoker loose the race
of putting data to operational data store, it fails if
it succeed.

Change-Id: Ic73c5b7bbe0f94b04bf932b67b872099f2c27b5c
Signed-off-by: Anil Vishnoi <vishnoianil@gmail.com>