Updated releng/autorelease Project: neutron  815bdc357638ddd644e1eea349070ea912651b0e
Bug 3165: PARTIAL fix for NeutronSecurityGroupInterface transcriber issue

This is a partial fix for regression introduced in
https://git.opendaylight.org/gerrit/#/c/19296/

The bug is that improper conversion from string to enumerator is
missing. Fix is similar to: https://git.opendaylight.org/gerrit/#/c/19633/

The bug also caused NPE when values in security group json are null. Removed
uneeded logic that built whole security rule an then threw it away.

Change-Id: Ib70894e8269fe19529cd3f3cfb3ee366ea240a6f
Signed-off-by: Flavio Fernandes <ffernand@redhat.com>