Updated releng/autorelease Project: lispflowmapping  9773792ee2c06a59c395cb1d26bd9479d51517fe
Switch mappingservice implementation to config subsystem

As a first step, the RPCs provided by the mappingservice implementation
are registered through the config subsystem. However, for the time being
we keep the Activator, not to break consumers (e.g., southbound or
northbound) that depend on it. As next steps, we'll switch the
consumers, one by one, to the config subsystem and once the process is
complete, we'll remove the Activator altogether.

Change-Id: Id8ff8549acf9769e22752dafd1aba326578bd0f5
Signed-off-by: Florin Coras <fcoras@cisco.com>